Split onboarding flow controller
